fix: PMD findings

+++ s4hana/s4hana-core/src/main/java/com/sap/cloud/sdk/s4hana/types/Year.java
@@ -46,7 +46,7 @@ public class Year extends Number implements Comparable<Year>
     @Override
     public long longValue()
     {
-        return year.longValue();
+        return year;
     }
 
     @Override
